+++ zcash_client_backend/src/wallet.rs
@@ -7,12 +7,10 @@ use zcash_primitives::{
     keys::OutgoingViewingKey,
     legacy::TransparentAddress,
     merkle_tree::IncrementalWitness,
-    sapling::{
-        note::ExtractedNoteCommitment, Diversifier, Node, Note, Nullifier, PaymentAddress, Rseed,
-    },
+    sapling,
     transaction::{
         components::{
-            sapling,
+            sapling::fees as sapling_fees,
             transparent::{self, OutPoint, TxOut},
             Amount,
         },
@@ -27,10 +25,8 @@ use zcash_primitives::{
 pub struct WalletTx<N> {
     pub txid: TxId,
     pub index: usize,
-    pub num_spends: usize,
-    pub num_outputs: usize,
-    pub shielded_spends: Vec<WalletShieldedSpend>,
-    pub shielded_outputs: Vec<WalletShieldedOutput<N>>,
+    pub sapling_spends: Vec<WalletSaplingSpend>,
+    pub sapling_outputs: Vec<WalletSaplingOutput<N>>,
 }
 
 #[derive(Debug, Clone)]
@@ -90,38 +86,107 @@ impl transparent::fees::InputView for WalletTransparentOutput {
 /// A subset of a [`SpendDescription`] relevant to wallets and light clients.
 ///
 /// [`SpendDescription`]: zcash_primitives::transaction::components::SpendDescription
-pub struct WalletShieldedSpend {
-    pub index: usize,
-    pub nf: Nullifier,
-    pub account: AccountId,
+pub struct WalletSaplingSpend {
+    index: usize,
+    nf: sapling::Nullifier,
+    account: AccountId,
+}
+
+impl WalletSaplingSpend {
+    pub fn from_parts(index: usize, nf: sapling::Nullifier, account: AccountId) -> Self {
+        Self { index, nf, account }
+    }
+
+    pub fn index(&self) -> usize {
+        self.index
+    }
+    pub fn nf(&self) -> &sapling::Nullifier {
+        &self.nf
+    }
+    pub fn account(&self) -> AccountId {
+        self.account
+    }
 }
 
 /// A subset of an [`OutputDescription`] relevant to wallets and light clients.
 ///
 /// [`OutputDescription`]: zcash_primitives::transaction::components::OutputDescription
-pub struct WalletShieldedOutput<N> {
-    pub index: usize,
-    pub cmu: ExtractedNoteCommitment,
-    pub ephemeral_key: EphemeralKeyBytes,
-    pub account: AccountId,
-    pub note: Note,
-    pub to: PaymentAddress,
-    pub is_change: bool,
-    pub witness: IncrementalWitness<Node>,
-    pub nf: N,
+pub struct WalletSaplingOutput<N> {
+    index: usize,
+    cmu: sapling::note::ExtractedNoteCommitment,
+    ephemeral_key: EphemeralKeyBytes,
+    account: AccountId,
+    note: sapling::Note,
+    is_change: bool,
+    witness: IncrementalWitness<sapling::Node>,
+    nf: N,
+}
+
+impl<N> WalletSaplingOutput<N> {
+    /// Constructs a new `WalletSaplingOutput` value from its constituent parts.
+    #[allow(clippy::too_many_arguments)]
+    pub fn from_parts(
+        index: usize,
+        cmu: sapling::note::ExtractedNoteCommitment,
+        ephemeral_key: EphemeralKeyBytes,
+        account: AccountId,
+        note: sapling::Note,
+        is_change: bool,
+        witness: IncrementalWitness<sapling::Node>,
+        nf: N,
+    ) -> Self {
+        Self {
+            index,
+            cmu,
+            ephemeral_key,
+            account,
+            note,
+            is_change,
+            witness,
+            nf,
+        }
+    }
+
+    pub fn index(&self) -> usize {
+        self.index
+    }
+    pub fn cmu(&self) -> &sapling::note::ExtractedNoteCommitment {
+        &self.cmu
+    }
+    pub fn ephemeral_key(&self) -> &EphemeralKeyBytes {
+        &self.ephemeral_key
+    }
+    pub fn account(&self) -> AccountId {
+        self.account
+    }
+    pub fn note(&self) -> &sapling::Note {
+        &self.note
+    }
+    pub fn is_change(&self) -> bool {
+        self.is_change
+    }
+    pub fn witness(&self) -> &IncrementalWitness<sapling::Node> {
+        &self.witness
+    }
+    pub fn witness_mut(&mut self) -> &mut IncrementalWitness<sapling::Node> {
+        &mut self.witness
+    }
+    pub fn nf(&self) -> &N {
+        &self.nf
+    }
 }
 
 /// Information about a note that is tracked by the wallet that is available for spending,
 /// with sufficient information for use in note selection.
 pub struct SpendableNote<NoteRef> {
     pub note_id: NoteRef,
-    pub diversifier: Diversifier,
+    pub diversifier: sapling::Diversifier,
     pub note_value: Amount,
-    pub rseed: Rseed,
-    pub witness: IncrementalWitness<Node>,
+    pub rseed: sapling::Rseed,
+    pub witness: IncrementalWitness<sapling::Node>,
 }
 
-impl<NoteRef> sapling::fees::InputView<NoteRef> for SpendableNote<NoteRef> {
+impl<NoteRef> sapling_fees::InputView<NoteRef> for SpendableNote<NoteRef> {
     fn note_id(&self) -> &NoteRef {
         &self.note_id
     }