+++ zcash_client_sqlite/src/wallet/transact.rs
@@ -62,10 +62,7 @@ fn to_spendable_note(row: &Row) -> Result<SpendableNote<NoteId>, SqliteClientErr
     })
 }
 
-#[deprecated(
-    note = "This method will be removed in a future update. Use zcash_client_backend::data_api::WalletRead::get_spendable_sapling_notes instead."
-)]
-pub fn get_spendable_sapling_notes<P>(
+pub(crate) fn get_spendable_sapling_notes<P>(
     wdb: &WalletDb<P>,
     account: AccountId,
     anchor_height: BlockHeight,
@@ -73,9 +70,9 @@ pub fn get_spendable_sapling_notes<P>(
 ) -> Result<Vec<SpendableNote<NoteId>>, SqliteClientError> {
     let mut stmt_select_notes = wdb.conn.prepare(
         "SELECT id_note, diversifier, value, rcm, witness
-            FROM received_notes
-            INNER JOIN transactions ON transactions.id_tx = received_notes.tx
-            INNER JOIN sapling_witnesses ON sapling_witnesses.note = received_notes.id_note
+            FROM sapling_received_notes
+            INNER JOIN transactions ON transactions.id_tx = sapling_received_notes.tx
+            INNER JOIN sapling_witnesses ON sapling_witnesses.note = sapling_received_notes.id_note
             WHERE account = :account
             AND spent IS NULL
             AND transactions.block <= :anchor_height
@@ -104,10 +101,7 @@ pub fn get_spendable_sapling_notes<P>(
     notes.collect::<Result<_, _>>()
 }
 
-#[deprecated(
-    note = "This method will be removed in a future update. Use zcash_client_backend::data_api::WalletRead::select_spendable_sapling_notes instead."
-)]
-pub fn select_spendable_sapling_notes<P>(
+pub(crate) fn select_spendable_sapling_notes<P>(
     wdb: &WalletDb<P>,
     account: AccountId,
     target_value: Amount,
@@ -138,8 +132,8 @@ pub fn select_spendable_sapling_notes<P>(
                 SELECT id_note, diversifier, value, rcm,
                     SUM(value) OVER
                         (PARTITION BY account, spent ORDER BY id_note) AS so_far
-                FROM received_notes
-                INNER JOIN transactions ON transactions.id_tx = received_notes.tx
+                FROM sapling_received_notes
+                INNER JOIN transactions ON transactions.id_tx = sapling_received_notes.tx
                 WHERE account = :account 
                 AND spent IS NULL 
                 AND transactions.block <= :anchor_height
@@ -360,7 +354,7 @@ mod tests {
                 available,
                 required
             })
-            if available == Amount::zero() && required == Amount::from_u64(1001).unwrap()
+            if available == Amount::zero() && required == Amount::from_u64(10001).unwrap()
         );
     }
 
@@ -443,7 +437,7 @@ mod tests {
                 required
             })
             if available == Amount::from_u64(50000).unwrap()
-                && required == Amount::from_u64(71000).unwrap()
+                && required == Amount::from_u64(80000).unwrap()
         );
 
         // Mine blocks SAPLING_ACTIVATION_HEIGHT + 2 to 9 until just before the second
@@ -478,7 +472,7 @@ mod tests {
                 required
             })
             if available == Amount::from_u64(50000).unwrap()
-                && required == Amount::from_u64(71000).unwrap()
+                && required == Amount::from_u64(80000).unwrap()
         );
 
         // Mine block 11 so that the second note becomes verified
@@ -574,7 +568,7 @@ mod tests {
                 available,
                 required
             })
-            if available == Amount::zero() && required == Amount::from_u64(3000).unwrap()
+            if available == Amount::zero() && required == Amount::from_u64(12000).unwrap()
         );
 
         // Mine blocks SAPLING_ACTIVATION_HEIGHT + 1 to 21 (that don't send us funds)
@@ -608,7 +602,7 @@ mod tests {
                 available,
                 required
             })
-            if available == Amount::zero() && required == Amount::from_u64(3000).unwrap()
+            if available == Amount::zero() && required == Amount::from_u64(12000).unwrap()
         );
 
         // Mine block SAPLING_ACTIVATION_HEIGHT + 22 so that the first transaction expires
@@ -758,7 +752,7 @@ mod tests {
         let dfvk = usk.sapling().to_diversifiable_full_viewing_key();
 
         // Add funds to the wallet in a single note
-        let value = Amount::from_u64(51000).unwrap();
+        let value = Amount::from_u64(60000).unwrap();
         let (cb, _) = fake_compact_block(
             sapling_activation_height(),
             BlockHash([0; 32]),
@@ -812,7 +806,7 @@ mod tests {
         let dfvk = usk.sapling().to_diversifiable_full_viewing_key();
 
         // Add funds to the wallet in a single note
-        let value = Amount::from_u64(51000).unwrap();
+        let value = Amount::from_u64(60000).unwrap();
         let (cb, _) = fake_compact_block(
             sapling_activation_height(),
             BlockHash([0; 32]),