+++ zcash_primitives/src/transaction/components/amount.rs
@@ -8,7 +8,7 @@ use orchard::value as orchard;
 pub const COIN: i64 = 1_0000_0000;
 pub const MAX_MONEY: i64 = 21_000_000 * COIN;
 
-pub const DEFAULT_FEE: Amount = Amount(1000);
+pub const DEFAULT_FEE: Amount = Amount(10_000);
 
 /// A type-safe representation of some quantity of Zcash.
 ///