+++ zcash_primitives/CHANGELOG.md
@@ -8,6 +8,8 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 ### Changed
 - Bumped dependencies to `secp256k1 0.26`, `hdwallet 0.4`.
+- `zcash_primitives::transactions::component::amount::DEFAULT_FEE` increased zip317 
+  minimum possible fee.
 
 ## [0.11.0] - 2023-04-15
 ### Added